+++ components/layout/grid/src/gridItem/index.js
@@ -1,12 +1,15 @@
 import {useMemo} from 'react'
 import PropTypes from 'prop-types'
 import cx from 'classnames'
+import PolymorphicElement from '@s-ui/react-primitive-polymorphic-element'
 
 import {BASE_CLASS, CELL_NUMBERS} from '../settings.js'
 import {getColSpanClassNamesTransform} from './settings.js'
 
 export default function LayoutGridItem({
+  as = 'div',
   children,
+  className,
   colSpan = 1,
   l,
   lOffset,
@@ -36,19 +39,29 @@ export default function LayoutGridItem({
     mOffset && `${BASE_CLASS}-item--mOffset-${mOffset}`,
     lOffset && `${BASE_CLASS}-item--lOffset-${lOffset}`,
     xlOffset && `${BASE_CLASS}-item--xlOffset-${xlOffset}`,
-    xxlOffset && `${BASE_CLASS}-item--xxlOffset-${xxlOffset}`
+    xxlOffset && `${BASE_CLASS}-item--xxlOffset-${xxlOffset}`,
+    className
   )
 
-  return <div className={classNames}>{children}</div>
+  return (
+    <PolymorphicElement as={as} className={classNames}>
+      {children}
+    </PolymorphicElement>
+  )
 }
 
 LayoutGridItem.displayName = 'LayoutGridItem'
 
 LayoutGridItem.propTypes = {
+  as: PropTypes.elementType,
   /**
    * The content of the component.
    */
   children: PropTypes.node,
+  /**
+   * Allows you to add extra styles and avoid extra DOM elements to style purposes.
+   */
+  className: PropTypes.string,
   /***
    * Defines the number of columns an item should span
    */